feature/hostname: move lazy_static var to lib.rs
